Extract MetadataDumper, implement error reporter, enforce conventions
- Created `src/MetadataDumper.ts` to encapsulate the core metadata chunking/processing logic, extracting it from the main `Plugin` class to adhere to the Single Responsibility Principle.
- Eliminated magic numbers by moving delay and interval values to well-named constants.
- Created `src/errorReporter.ts` to funnel unexpected errors into a centralized log, satisfying the robust error handling directive.
- Caught formerly unhandled interval and command promise rejections in `main.ts` and funneled them to the error reporter.
- Codified project guidelines and layout structure in an `AGENTS.md` file to instruct future agents.
- Enforced tool versioning via `mise.toml`.
